requirements.txt
absl-py
contourpy
cycler
filelock
fonttools
fsspec
grpcio
Jinja2
joblib
kiwisolver
Markdown
MarkupSafe
matplotlib
mpmath
networkx
numpy
nvidia-cublas-cu12
nvidia-cuda-cupti-cu12
nvidia-cuda-nvrtc-cu12
nvidia-cuda-runtime-cu12
nvidia-cudnn-cu12
nvidia-cufft-cu12
nvidia-cufile-cu12
nvidia-curand-cu12
nvidia-cusolver-cu12
nvidia-cusparse-cu12
nvidia-cusparselt-cu12
nvidia-nccl-cu12
nvidia-nvjitlink-cu12
nvidia-nvtx-cu12
opencv-python
packaging
pillow
protobuf
pyparsing
python-dateutil
PyYAML
scikit-learn
scipy
setuptools
six
sympy
tensorboard
tensorboard-data-server
threadpoolctl
torch
torchvision
tqdm
triton
typing_extensions
Werkzeug
einops
timm
# MMDetection dependencies (for CFINet object detection)
# Note: CFINet should be installed separately from the CFI/CFINet directory
# Run: cd CFI/CFINet && pip install -v -e .
mmcv-full>=1.5.0
mmdet>=2.26.0